Support multiple services, validate session
Refactor to dynamically select transcription service based on query parameter. Added session and token retrieval to enhance request validation. Improved error handling and cleanup logic for temporary files.

+++ app/api/v2/file/[id]/transcribe/route.ts
@@ -8,14 +8,30 @@ import {authOptions} from '@/pages/api/auth/[...nextauth]';
 import {tmpdir} from 'os';
 import {join} from 'path';
 import fs from "fs";
-import TranscriptionService from "@/services/transcriptionService";
+import {TranscriptionServiceFactory} from "@/services/transcriptionService";
 import {promisify} from "util";
 
 const unlinkAsync = promisify(fs.unlink);
 
 export async function GET(request: NextRequest, { params }: { params: { id: string } }) {
+  // @ts-ignore
+  const token: JWT | null = await getToken({ req: request });
+  if (!token)
+    throw new Error(`Token could not be pulled from request: ${request}`);
+
+  const session: Session | null = await getServerSession(authOptions as any);
+  if (!session) throw new Error("Failed to pull session!");
 
   const { id } = params;
+
+  const { searchParams } = new URL(request.url);
+  let transcriptionServiceName = searchParams.get('service');
+  if (!transcriptionServiceName || (!['whisper', ''].includes(transcriptionServiceName))) {
+    transcriptionServiceName = 'azureCognitiveSpeechService'
+  }
+
+
+  let transcript: string | undefined;
 
   try {
     // Get the session and user ID
@@ -53,21 +69,27 @@ export async function GET(request: NextRequest, { params }: { params: { id: stri
     const tmpFilePath = join(tmpdir(), `${Date.now()}_${id}`);
     await blockBlobClient.downloadToFile(tmpFilePath);
 
-    // Initialize the TranscriptionService
-    const transcriptionService = new TranscriptionService();
+    // @ts-expect-error This is handled in logic above making sure that the name is not null and valid
+    const transcriptionService = TranscriptionServiceFactory.getTranscriptionService(transcriptionServiceName);
 
     // Transcribe the file
-    const transcript = await transcriptionService.transcribe(tmpFilePath);
+    transcript = await transcriptionService.transcribe(tmpFilePath);
 
     // Clean up temporary files
     await unlinkAsync(tmpFilePath);
 
     // Clean up temporary files
-    await fs.unlink(tmpFilePath, ()=>null);
+    fs.exists(tmpFilePath, (exists: boolean) => {
+        if (exists) {
+          fs.unlink(tmpFilePath, ()=>null);
+        }
+    })
 
     return NextResponse.json({ transcript });
   } catch (error) {
     console.error('Error during transcription:', error);
+    if (transcript)
+      return NextResponse.json({ transcript });
     return NextResponse.json({ error: 'Failed to transcribe file' }, { status: 500 });
   }
   // return NextResponse.json({ error: 'Failed to transcribe file' }, { status: 500 });
